requirements.txt
flask
pywebview
PyPDF2
pycryptodome
pywin32
requests
pillow
pyinstaller 